src/test/java/com/example/demo/controlador/UserControllerTest.java
package com.example.demo.controlador;

import static org.mockito.ArgumentMatchers.any;
import static org.mockito.ArgumentMatchers.anyLong;
import static org.mockito.ArgumentMatchers.anyString;
import static org.mockito.Mockito.mock;
import static org.mockito.Mockito.when;
import static org.springframework.security.test.web.servlet.request.SecurityMockMvcRequestPostProcessors.csrf;
import static org.springframework.security.test.web.servlet.setup.SecurityMockMvcConfigurers.springSecurity;
import static org.springframework.test.web.servlet.request.MockMvcRequestBuilders.*;
import static org.springframework.test.web.servlet.result.MockMvcResultMatchers.*;

import java.util.Collections;

import org.junit.jupiter.api.BeforeEach;
import org.junit.jupiter.api.Test;
import org.mockito.InjectMocks;
import org.mockito.Mock;
import org.springframework.beans.factory.annotation.Autowired;
import org.springframework.boot.test.autoconfigure.web.servlet.WebMvcTest;
import org.springframework.boot.test.mock.mockito.MockBean;
import org.springframework.data.domain.Page;
import org.springframework.data.domain.PageImpl;
import org.springframework.data.domain.PageRequest;
import org.springframework.security.core.Authentication;
import org.springframework.security.core.context.SecurityContext;
import org.springframework.security.core.context.SecurityContextHolder;
import org.springframework.security.test.context.support.WithMockUser;
import org.springframework.test.web.servlet.MockMvc;
import org.springframework.test.web.servlet.setup.MockMvcBuilders;
import org.springframework.web.context.WebApplicationContext;

import com.example.demo.dto.UsuarioDTO;
import com.example.demo.entidad.Comentario;
import com.example.demo.entidad.Usuario;
import com.example.demo.servicio.comentario.ComentarioServicio;
import com.example.demo.servicio.usuario.UsuarioServicio;

@WebMvcTest(UserController.class)
public class UserControllerTest {

	@Autowired
	private MockMvc mockMvc;

	@MockBean
	private ComentarioServicio comentarioServicio;

	@MockBean
	private UsuarioServicio usuarioServicio;

	@InjectMocks
	private UserController userController;

	@Autowired
	private WebApplicationContext webApplicationContext;

	@BeforeEach
	public void setup() {
		
		/**
		 *  Construir una instancia de MockMvc. 
		 *  Esto es crucial para simular peticiones HTTP 
		 *  a tu controlador en un entorno de prueba aislado.
		 */
		mockMvc = MockMvcBuilders.webAppContextSetup(webApplicationContext)
				.apply(springSecurity()) // Asegúrate de aplicar la seguridad de Spring
				.build();

	    /** 
	     * Configurar un objeto Page<Comentario> no nulo para evitar NullPointerException
	     */
		Page<Comentario> comentariosPage = new PageImpl<>(Collections.emptyList(), PageRequest.of(0, 10), 0);
		when(comentarioServicio.listarTodos(any(PageRequest.class))).thenReturn(comentariosPage);
		/**
		 * UsuarioDTO mockeado 
		 * para ser devuelto por usuarioServicio.obtenerUsuarioDTO.
		 */
		 UsuarioDTO usuariodtoMock = new UsuarioDTO();
		 usuariodtoMock.setUsername("user");
		 usuariodtoMock.setId(1L);
		    // ... set other properties of usuarioMock as needed
		    when(usuarioServicio.obtenerUsuarioDTO(anyString())).thenReturn(usuariodtoMock);
		
		    // Configura el comportamiento esperado para el servicio de usuario.
		    Usuario usuarioMock = new Usuario();
		    usuarioMock.setId(1L); // Suponiendo que el ID del usuario es 1
		    when(usuarioServicio.obtenerPorId(anyLong())).thenReturn(usuarioMock);  
		    
		/**
		 * Simular un contexto de seguridad con un usuario autenticado. 
		 * Esto es esencial para pruebas que dependen de la autenticación 
		 * y autorización de Spring Security.
		 */
		Authentication authentication = mock(Authentication.class);
		when(authentication.getName()).thenReturn("user");
		SecurityContext securityContext = mock(SecurityContext.class);
		when(securityContext.getAuthentication()).thenReturn(authentication);
		SecurityContextHolder.setContext(securityContext);
	}

	@Test
	@WithMockUser(username = "user", roles = "USER")
	public void testUserHomePage() throws Exception {
		
		mockMvc
		.perform(get("/user/home")).andExpect(status().isOk())
        .andExpect(model().attributeExists("username"))
        .andExpect(model().attributeExists("usuarioDTO"))
        .andExpect(model().attributeExists("comentarios"))
     // Agrega más aserciones para verificar los atributos del modelo si es necesario
		.andExpect(view().name("auth/user/home"));
		
	}
	@Test
	@WithMockUser(username = "user", roles = "USER")
	public void testMostrarFormularioComentario() throws Exception {
	    mockMvc.perform(get("/user/agregarComentario"))
	           .andExpect(status().isOk())
	           .andExpect(view().name("/auth/user/formCrearComentario"))
	           .andExpect(model().attributeExists("comentario"))
	           // Aquí puedes añadir más aserciones si el método añade más atributos al modelo
	           ;
	}
	
	@Test
	@WithMockUser(username = "user", roles = "USER")
	public void testAgregarComentario() throws Exception {
		// Simula una solicitud POST al controlador
	    mockMvc.perform(post("/user/agregarComentario")
	                    .param("texto", "Este es un comentario de prueba")
	                    .param("usuarioId", "1") // Asegúrate de que estos parámetros sean los esperados por el controlador
	                    .with(csrf())) // Importante para formularios POST
	           .andExpect(status().isOk());
	          
      
	}

}
